# src/OpenMetrc.Scraper/OpenApi/metrc-swagger-v2.yaml
swagger: 2.0
info:
  title: METRC API
  version: v2
paths:
...
  /strains/v2/{id}:
    delete:
      tags:
      - Strains
      summary: ''
      description: '<b>Permissions Required</b>: Manage Strains'
      externalDocs:
        x-doc-key: delete_strains_v2_{id}
      operationId: DeleteStrainById
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which Strain to delete.
        format: ''
      responses:
        200:
          description: Ok
    get:
      tags:
      - Strains
      summary: ''
      description: '<b>Permissions Required</b>: Manage Strains'
      externalDocs:
        x-doc-key: get_strains_v2_{id}
      operationId: GetStrainById
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        description: If specified, the Strain will be validated against the specified License Number. If not specified, the Strain will be validated against all of the User's current Facilities. Please note that if the Strain is not valid for the specified License Number, a 401 Unauthorized status will be returned.
        format: ''
      responses:
        200:
          description: Ok
  /strains/v2/active:
    get:
      tags:
      - Strains
      summary: ''
      description: '<b>Permissions Required</b>: Manage Strains'
      externalDocs:
        x-doc-key: get_strains_v2_active
      operationId: GetStrainActive
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to return the list of active strains.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      - type: string
        name: lastModifiedStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: lastModifiedEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
  /strains/v2/inactive:
    get:
      tags:
      - Strains
      summary: ''
      description: '<b>Permissions Required</b>: Manage Strains'
      externalDocs:
        x-doc-key: get_strains_v2_inactive
      operationId: GetStrainInactive
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to return the list of inactive Strains.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      responses:
        200:
          description: Ok
  /strains/v2/:
    post:
      tags:
      - Strains
      summary: ''
      description: '<b>Permissions Required</b>: Manage Strains'
      externalDocs:
        x-doc-key: post_strains_v2_
      operationId: PostStrains
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to record the list of Strains.
        format: ''
      responses:
        200:
          description: Ok
    put:
      tags:
      - Strains
      summary: ''
      description: '<b>Permissions Required</b>: Manage Strains'
      externalDocs:
        x-doc-key: put_strains_v2_
      operationId: PutStrains
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The License Number of the Facility for which to update the list of Strains.
        format: ''
      responses:
        200:
          description: Ok
...